backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pg8000==1.31.2

# Authentication & Security
PyJWT==2.10.1
cryptography==43.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation
pydantic==2.10.3
email-validator==2.2.0
python-multipart==0.0.17

# API & Middleware
slowapi==0.1.9
pydantic-settings==2.7.0

# Monitoring
sentry-sdk==2.19.2
python-dotenv==1.0.1

# External Services
stripe==11.3.0
openai==1.58.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.1
httpx==0.28.1

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0
